refactor : refreshKey 하드 코딩 제거 (#23)
